add history when percents calc

+++ code/migrations/m180615_175804_deposit_history.php
@@ -22,7 +22,7 @@ class m180615_175804_deposit_history extends Migration
             'client_id' => $this->integer()->unsigned(),
             'deposit_id' => $this->integer()->unsigned(),
             'type' => $this->tinyInteger(1)->unsigned(),
-            'value' => $this->integer()->unsigned(),
+            'value' => $this->bigInteger()->unsigned(),
             'slice' => $this->integer()->unsigned(),
             'created_at' => $this->integer()->unsigned()
         ], $tableOptions);
